+++ src/platform/packages/private/kbn-reporting/public/share/index.ts
@@ -7,11 +7,9 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export {
-  reportingPDFExportProvider,
-  reportingPNGExportProvider,
-} from './share_context_menu/register_pdf_png_modal_reporting';
-export { reportingCsvExportProvider } from './share_context_menu/register_csv_modal_reporting';
+export { reportingCsvExportShareIntegration } from './integrations/csv';
+export { reportingPDFExportShareIntegration } from './integrations/pdf';
+export { reportingPNGExportShareIntegration } from './integrations/png';
 export type { JobParamsProviderOptions, StartServices } from './share_context_menu';
 export { getSharedComponents } from './shared';
 export type { ReportingPublicComponents } from './shared';
