+++ src/platform/packages/private/kbn-reporting/public/share/integrations/csv/csv_export_config.tsx
@@ -7,18 +7,19 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import { i18n } from '@kbn/i18n';
-import { toMountPoint } from '@kbn/react-kibana-mount';
 import React from 'react';
 import { firstValueFrom } from 'rxjs';
+import { i18n } from '@kbn/i18n';
+import { toMountPoint } from '@kbn/react-kibana-mount';
 import type { SerializedSearchSourceFields } from '@kbn/data-plugin/common';
-import { FormattedMessage, InjectedIntl } from '@kbn/i18n-react';
-import { ShareContext, type ExportShare } from '@kbn/share-plugin/public';
-import { LocatorParams } from '@kbn/reporting-common/types';
-import { ReportParamsGetter, ReportParamsGetterOptions } from '../../types';
-import { getSearchCsvJobParams, CsvSearchModeParams } from '../shared/get_search_csv_job_params';
-import type { ExportModalShareOpts } from '.';
-import { checkLicense } from '../..';
+import type { InjectedIntl } from '@kbn/i18n-react';
+import { FormattedMessage } from '@kbn/i18n-react';
+import type { ShareContext, ExportShare } from '@kbn/share-plugin/public';
+import type { LocatorParams } from '@kbn/reporting-common/types';
+import type { ReportParamsGetter, ReportParamsGetterOptions } from '../../../types';
+import type { CsvSearchModeParams } from '../../shared/get_search_csv_job_params';
+import { getSearchCsvJobParams } from '../../shared/get_search_csv_job_params';
+import type { ExportModalShareOpts } from '../../share_context_menu';
 
 export const getCsvReportParams: ReportParamsGetter<
   ReportParamsGetterOptions & { forShareUrl?: boolean },
@@ -51,14 +52,15 @@ export const getCsvReportParams: ReportParamsGetter<
   };
 };
 
-export const reportingCsvExportProvider = ({
-  apiClient,
-  startServices$,
-}: ExportModalShareOpts): ExportShare => {
-  const getShareMenuItems = ({
+/**
+ * @description Returns config for the CSV export integration
+ */
+export const getShareMenuItems =
+  ({ apiClient, startServices$ }: ExportModalShareOpts) =>
+  ({
     objectType,
     sharingData,
-  }: ShareContext): ReturnType<ExportShare['config']> => {
+  }: ShareContext): ReturnType<ExportShare['config']> extends Promise<infer R> ? R : never => {
     const getSearchModeParams = (forShareUrl?: boolean): CsvSearchModeParams =>
       getCsvReportParams({ sharingData, forShareUrl });
 
@@ -168,30 +170,3 @@ export const reportingCsvExportProvider = ({
       },
     };
   };
-
-  return {
-    shareType: 'integration',
-    id: 'csvReports',
-    groupId: 'export',
-    config: getShareMenuItems,
-    prerequisiteCheck: ({ license, capabilities }) => {
-      if (!license) {
-        return false;
-      }
-
-      const licenseCheck = checkLicense(license.check('reporting', 'basic'));
-
-      const licenseHasCsvReporting = licenseCheck.showLinks;
-
-      const capabilityHasCsvReporting =
-        capabilities.discover_v2?.generateCsv === true ||
-        capabilities.reportingLegacy?.generateReport === true;
-
-      if (!(licenseHasCsvReporting && capabilityHasCsvReporting)) {
-        return false;
-      }
-
-      return true;
-    },
-  };
-};
