+++ src/platform/plugins/shared/discover/public/application/main/hooks/use_has_share_integration.ts
@@ -6,16 +6,14 @@
  * your election, the "Elastic License 2.0", the "GNU Affero General Public
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
-import type { Capabilities } from '@kbn/core/public';
-export { getSharedComponents } from './get_shared_components';
-export type { ApplicationProps, ReportingPublicComponents } from './get_shared_components';
 
-export const hasCapabilityByKey = (
-  capabilities: Capabilities,
-  capabilityKey: keyof Capabilities
-) => {
-  return (
-    capabilities[capabilityKey]?.generateScreenshot === true ||
-    capabilities.reportingLegacy?.generateReport === true
-  );
-};
+import { useMemo } from 'react';
+import type { DiscoverServices } from '../../../build_services';
+
+export function useHasShareIntegration({ share }: DiscoverServices) {
+  return useMemo(() => {
+    if (!share) return false;
+
+    return share.availableIntegrations('search', 'export').length > 0;
+  }, [share]);
+}
