+++ src/platform/plugins/shared/share/public/index.ts
@@ -27,7 +27,9 @@ export type {
   ShareContextMenuPanelItem,
   BrowserUrlService,
   ExportShare,
+  ExportShareConfig,
   ExportShareDerivatives,
+  RegisterShareIntegrationArgs,
 } from './types';
 
 export type { RedirectOptions } from '../common/url_service';
