+++ x-pack/platform/plugins/shared/lens/public/plugin.ts
@@ -135,13 +135,9 @@ import { setupExpressions } from './expressions';
 import { OpenInDiscoverDrilldown } from './trigger_actions/open_in_discover_drilldown';
 import { ChartInfoApi } from './chart_info_api';
 import { type LensAppLocator, LensAppLocatorDefinition } from '../common/locator/locator';
-import { downloadCsvLensShareProvider } from './app_plugin/csv_download_provider/csv_download_provider';
-import { LensDocument } from './persistence/saved_object_store';
-import {
-  CONTENT_ID,
-  LATEST_VERSION,
-  LensSavedObjectAttributes,
-} from '../common/content_management';
+import type { LensDocument } from './persistence';
+import type { LensSavedObjectAttributes } from '../common/content_management';
+import { CONTENT_ID, LATEST_VERSION } from '../common/content_management';
 import type { EditLensConfigurationProps } from './app_plugin/shared/edit_on_the_fly/get_edit_lens_configuration';
 import { LensRenderer } from './react_embeddable/renderer/lens_custom_renderer_component';
 import {
@@ -149,6 +145,7 @@ import {
   ACTION_EDIT_LENS_EMBEDDABLE,
   IN_APP_EMBEDDABLE_EDIT_TRIGGER,
 } from './trigger_actions/open_lens_config/constants';
+import { downloadCsvLensShareProvider } from './app_plugin/csv_download_provider/csv_download_provider';
 
 export type { SaveProps } from './app_plugin';
